pyproject.toml
[project]
name = "agentic-rag-with-inngest-llamaindex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.20.0",
    "fastapi>=0.124.4",
    "groq>=0.37.1",
    "inngest>=0.5.13",
    "llama-index-core>=0.14.10",
    "llama-index-readers-file>=0.5.5",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "streamlit>=1.52.1",
    "uvicorn>=0.38.0",
]
